Populate repo from legacy folder.
Fortran program opac.f90 tested, builds and runs with no issues.

Partial documentation in Help/*.html, should be converted to .md
at some point, and edited. Basic usage is simple enough and can be
found in the opac.f90 header. Template configuration and atmosphere
files are found in the same dirctory as the source. The main
program is controled by reading cfg.dat, the name is hardcoded
in opac.f90. The configuration parameters are documented in the
configuration file itself.